requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic<2.0
python-jose[cryptography]
passlib[bcrypt]
python-multipart
httpx
